restructure the folders
