--- simple_greet/hello.py
@@ -1,3 +1,0 @@
-from greeting import greet
-
-greet("Simon")
