+++ src/simple_greet/hello.py
@@ -1,0 +1,3 @@
+from src.simple_greet.greeting import greet
+
+greet("Simon")
